Fixed some bugs and implemented Incidence.delete

+++ src/de/uni_koblenz/jgralab/impl/disk/BinaryEdgeImpl.java
@@ -14,12 +14,12 @@ public abstract class BinaryEdgeImpl extends EdgeImpl implements BinaryEdge {
 	 */
 	private static final long serialVersionUID = -8155387019211417812L;
 
-	protected BinaryEdgeImpl(int anId, GraphDatabaseBaseImpl graphDatabase)
+	protected BinaryEdgeImpl(long anId, GraphDatabaseBaseImpl graphDatabase)
 			throws IOException {
 		super(anId, graphDatabase);
 	}
 
-	protected BinaryEdgeImpl(int anId, GraphDatabaseBaseImpl graphDatabase,
+	protected BinaryEdgeImpl(long anId, GraphDatabaseBaseImpl graphDatabase,
 			EdgeContainer container) throws IOException {
 		super(anId, graphDatabase, container);
 	}
